requirements.txt
python-nfsn==1.1.1
PyYAML==5.3.1